fix: set version of pyspark setup.py (#77)
Previously, we specified a specific version of pyspark in requirements.txt, but not in setup.py.  This can cause inconsistencies in installation depending on whether this package is installed in a virtualenv from source versus "pip installed" from PyPI.

To fix this issue and provide some consistency, have setup.py depend on at least minimally the version of pyspark pinned in requirements.txt.